feature: ucelfka splitted into two creates
* ucelofka - builds a binary
* ucelofka-data - contains data definition
