+++ ucelofka-data/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ucelofka-data"
+version = "0.1.0"
+authors = ["Stepan Henek"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "~1.0.26"
+chrono = "~0.4.10"
+serde = { version = "~1.0.104", features = ["derive"]}
+serde_yaml = "~0.8.11"
