+++ ucelofka-data/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod account;
+pub mod customer;
+pub mod entry;
+pub mod identity;
+pub mod invoice;
+pub mod template;
