+++ ucelofka/Cargo.toml
@@ -30,6 +30,7 @@ serde_json = "1"
 tera = "~1.5.0"
 tokio = { version = "1", features = ["rt", "rt-multi-thread"]}
 unic-langid = { version = "~0.9.0", features = ["macros"]}
+ucelofka-data = { path="../ucelofka-data/" }
 
 [dev-dependencies]
 assert_cmd = "~1.0.1"
